--- biff/__init__.py
@@ -1,1 +1,0 @@
-from .biff import DiffusionModel
